Update other branch checklists without swapping
You can now use <cmd> [-b <name> | --branch <name>] to make changes to a
different branch without switching first. Particularly handy if you're
deep in the middle of something and you have to add a new TODO to a
different branch.

It doesn't do any sanity check on the branch name; it is possible to
suggest a branch which doesn't exist. If you tried to enter:

    $ git todo add -b master Increment version number for beta

but it came out as `-b msater` you'd get a new checklist in
`.git/checklist/msater`. This is a blessing and a curse since you can
have "detached checklists" which live independently from any known
branch. But you can also add an item to a branch and end up losing it
due to a typo.