Remove branch names before stats lines
It was causing the branch names to be printed twice when you entered
`stats --all` which looked stupid. In the normal case the branch name
probably isn't needed (or maybe it should be required for checklists
too?). Anyway it's gone for now.

Also corrected a couple of related issues when stats were or weren't
getting new lines printed depending on whether there was anything there.